+++ src/components/ClassInstances/ImportCSVInstances.js
@@ -1,35 +1,30 @@
 import { useState } from 'react';
-import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap';
+import { Button, Modal, ModalHeader, ModalBody, ModalFooter, InputGroup } from 'reactstrap';
 import { createLiteralStatement } from 'services/backend/statements';
 import { createLiteral } from 'services/backend/literals';
 import { createResource } from 'services/backend/resources';
 import { PREDICATES } from 'constants/graphSettings';
-import CSVReader from 'react-csv-reader';
+import { useCSVReader } from 'react-papaparse';
 import { toast } from 'react-toastify';
 import PropTypes from 'prop-types';
 
 export default function ImportCSVInstances(props) {
     const [data, setData] = useState([]);
-    const [error, seError] = useState(null);
     const [isImporting, setIsImporting] = useState(false);
+    const { CSVReader } = useCSVReader();
 
-    const handleOnFileLoad = data => {
+    const handleOnFileLoad = ({ _data }) => {
         // Check the csv file
-        if (!data || data.length === 0 || data[0].length !== 2 || data[0][0].toLowerCase() !== 'label' || data[0][1].toLowerCase() !== 'uri') {
+        if (!_data || _data.length === 0 || _data[0].length !== 2 || _data[0][0].toLowerCase() !== 'label' || _data[0][1].toLowerCase() !== 'uri') {
             toast.error('Please upload a CSV file that has only two columns: Label and URI');
             setData([]);
         } else {
-            setData(data);
+            setData(_data);
         }
     };
 
-    const handleOnError = err => {
-        seError(err);
-    };
-
-    const csvReaderOptions = {
+    const PARSER_OPTIONS = {
         header: false,
-        dynamicTyping: true,
         skipEmptyLines: true,
     };
 
@@ -53,9 +48,10 @@ export default function ImportCSVInstances(props) {
                         props.callBack(); // Basically to refresh the list of instances
                         setIsImporting(false);
                         setData([]);
+                        props.toggle();
                     });
                 })
-                .catch(e => {
+                .catch(() => {
                     toast.error('Something went wrong when importing instances');
                     setIsImporting(false);
                     setData([]);
@@ -69,21 +65,35 @@ export default function ImportCSVInstances(props) {
             <ModalBody>
                 Please import your CSV file that has two columns: <b>Label</b> and <b>URI</b> using the form below.
                 <div className="mt-3">
-                    <CSVReader
-                        cssClass="csv-reader-input"
-                        label="Select a CSV File"
-                        onFileLoaded={handleOnFileLoad}
-                        onError={handleOnError}
-                        parserOptions={csvReaderOptions}
-                        inputStyle={{ marginLeft: '5px' }}
-                    />
+                    <CSVReader accept=".csv, text/csv" config={PARSER_OPTIONS} onUploadAccepted={result => handleOnFileLoad({ _data: result.data })}>
+                        {({ getRootProps, acceptedFile, ProgressBar }) => (
+                            <>
+                                <InputGroup>
+                                    <Button {...getRootProps()}>Browse file</Button>
+                                    <div
+                                        {...getRootProps()}
+                                        style={{
+                                            border: '1px solid #ccc',
+                                            lineHeight: 2.2,
+                                            paddingLeft: 10,
+                                            flexGrow: '1',
+                                        }}
+                                    >
+                                        {acceptedFile && acceptedFile.name}
+                                        {!acceptedFile && 'No file is selected'}
+                                    </div>
+                                </InputGroup>
+                                <ProgressBar style={{ backgroundColor: '#dbdde5' }} />
+                            </>
+                        )}
+                    </CSVReader>
                 </div>
             </ModalBody>
             <ModalFooter>
                 <Button color="light" onClick={props.toggle}>
                     Cancel
                 </Button>
-                <Button color="primary" onClick={handleImport} disabled={Boolean(isImporting || !data || data.length === 0 || error)}>
+                <Button color="primary" onClick={handleImport} disabled={Boolean(isImporting || !data || data.length === 0)}>
                     {!isImporting && <>Import {data.length > 2 ? ` ${data.length - 1} ` : ''} instances</>}
                     {isImporting && 'Importing ....'}
                 </Button>{' '}