+++ src/components/Review/MarkdownRenderer.js
@@ -1,7 +1,7 @@
 import { createReference, setUsedReferences as setUsedReferencesAction } from 'slices/reviewSlice';
 import { sanitize } from 'dompurify';
 import PropTypes from 'prop-types';
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import * as Showdown from 'showdown';
 import footnotes from 'showdown-footnotes';
@@ -23,6 +23,10 @@ const MarkdownRenderer = ({ text, id }) => {
     const referenceRegex = useMemo(() => /\[@(.*?)\]/gi, []);
     const dispatch = useDispatch();
 
+    const formatReferenceKey = useCallback(key => key.slice(0, -1).slice(2, key.length), []);
+
+    const getReferenceByKey = useCallback(key => references.find(reference => reference?.parsedReference?.id === key), [references]);
+
     const inlineReferences = {
         type: 'lang',
         regex: referenceRegex,
@@ -38,10 +42,6 @@ const MarkdownRenderer = ({ text, id }) => {
         },
     };
 
-    const formatReferenceKey = useCallback(key => key.slice(0, -1).slice(2, key.length), []);
-
-    const getReferenceByKey = useCallback(key => references.find(reference => reference?.parsedReference?.id === key), [references]);
-
     useEffect(() => {
         if (!text) {
             return;
@@ -49,8 +49,7 @@ const MarkdownRenderer = ({ text, id }) => {
         const _usedReferences = {};
         const matches = text.match(referenceRegex);
 
-        matches &&
-            matches.length &&
+        if (matches && matches.length) {
             matches.map(async key => {
                 const keyFormatted = formatReferenceKey(key);
                 const reference = getReferenceByKey(keyFormatted);
@@ -79,6 +78,7 @@ const MarkdownRenderer = ({ text, id }) => {
                 }
                 return null;
             });
+        }
         dispatch(setUsedReferencesAction({ references: _usedReferences, sectionId: id }));
     }, [text, references, referenceRegex, getReferenceByKey, dispatch, id, formatReferenceKey, fetchedDois, contributionId]);