+++ src/services/orkgNlp/index.js
@@ -116,3 +116,29 @@ export const saveFeedback = async ({ request, response, serviceName }) =>
 
 export const semantifyBioassays = text =>
     submitPostRequest(`${nlpServiceUrl}clustering/bioassays/`, { 'Content-Type': 'application/json' }, { text });
+
+export const extractTable = form =>
+    fetch(`${nlpServiceUrl}tools/pdf/table/extract`, {
+        method: 'POST',
+        body: form,
+    });
+
+export const convertPdf = form =>
+    fetch(`${nlpServiceUrl}tools/pdf/convert`, {
+        method: 'POST',
+        body: form,
+    });
+
+export const getRecommendedPredicates = async ({ title, abstract }) => {
+    const { payload } = await submitPostRequest(
+        `${nlpServiceUrl}clustering/predicates`,
+        {
+            'Content-Type': 'application/json',
+        },
+        {
+            title,
+            abstract,
+        },
+    );
+    return payload;
+};