+++ src/slices/reviewSlice.js
@@ -77,8 +77,10 @@ export const reviewSlice = createSlice({
         },
         sectionLinkUpdated: (state, { payload: { id, objectId, label } }) => {
             const index = state.sections.findIndex(section => section?.id === id);
-            state.sections[index].contentLink.objectId = objectId;
-            state.sections[index].contentLink.label = label;
+            state.sections[index].contentLink = {
+                objectId,
+                label,
+            };
         },
         sectionTypeUpdated: (state, { payload: { sectionId, sectionType } }) => {
             const index = state.sections.findIndex(section => section.id === sectionId);
@@ -168,8 +170,8 @@ export const reviewSlice = createSlice({
             state.usedReferences[sectionId] = references;
         },
     },
-    extraReducers: {
-        [LOCATION_CHANGE]: (state, { payload }) => {
+    extraReducers: builder => {
+        builder.addCase(LOCATION_CHANGE, (state, { payload }) => {
             const matchReview = match(ROUTES.REVIEW);
             const parsed_payload = matchReview(payload.location.pathname);
             if (parsed_payload && parsed_payload.params?.id === state.articleId) {
@@ -177,7 +179,7 @@ export const reviewSlice = createSlice({
                 return state;
             }
             return initialState;
-        },
+        });
     },
 });